TEDOPA

TEDOPA is a package for computing the chain mapping of spectral densities of both bosonic and fermionic environments. Starting from the spectral density function, the temperature and the chemical potential (if applicable), it computes the chain coefficients of the TEDOPA-transformed environment, applying a thermalization procedure such as T-TEDOPA or the thermofield transformation.

Installation

The package must be installed directly from GitHub, as it is not in any public registry. From the Julia REPL, type ] to enter the Pkg REPL mode and run:

add "https://github.com/phaerrax/TEDOPA.jl.git"

How to describe environments

Information about an environment (or a collections of environments) must be encoded in a JSON dictionary with the following structure:

{
    "environment": {
        "spectral_density_parameters": [1, 0.5],
        "spectral_density_function": "1/(10pi) * sqrt((2a[2]-a[1]+x)*(2a[2]+a[1]-x))",
        "domain": [0, 2],
        "temperature": 0.4,
        "chemical_potential": 0.5
    },
    "chain_length": 200,
    "PolyChaos_nquad": 5000
}

Each element of the "environments" subdictionary defines an environment as follows:

"spectral_density_function" is a string containing a valid Julia function of the variable x and possibly a set of parameters a[1], a[2], etc.;

"spectral_density_parameters" is a list of numbers that will replace a[1], a[2] and so on in the spectral density function;

"domain" is the domain of the spectral density function;

"temperature" is the temperature of the environment;

"chemical_potential" is the chemical potential of the environment.

There are then two additional parameters:

"chain_length" specifies the lengths of the resulting chains;

"PolyChaos_nquad" is a parameter for the numerical integration used to calculate the coefficients (usually a higher number means a higher precision).

How to compute the chain coefficients

This package offers several ways of computing the chain mapping, all derived from the original TEDOPA algorithm.

chainmapping_tedopa: standard chain mapping of a single environment;

chainmapping_ttedopa: thermalized chain mapping for bosonic environments;

chainmapping_tftedopa: single-chain thermalized chain mapping for fermionic environments;

chainmapping_thermofield: a thermofield transformation and then a chain mapping of the (fermionic) environment; it can also merge multiple multiple environments together (in this case, the "environment" subdictionary is replaced by a list of dictionaries, such as in the following example).

{
    "environment": [
        {
            "spectral_density_parameters": [],
            "spectral_density_function": "1/(pi*200)",
            "domain": [-100, 100],
            "temperature": 0,
            "chemical_potential": 0.01
        },
        {
            "spectral_density_parameters": [],
            "spectral_density_function": "1/(pi*200)",
            "domain": [-100, 100],
            "temperature": 0,
            "chemical_potential": -0.01
        }
    ],
    "chain_length": 200,
    "PolyChaos_nquad": 5000
}

Once you have a properly formatted dictionary in a JSON file, you can call one of the above functions by providing a file object or a string containing the file name; you can also provide the dictionary itself, directly.

Chain-mapping functions

Core algorithm

TEDOPA.chainmapping — Function

chainmapping(J::Function, support, L::Int; kwargs...)

Compute the first L coefficients of the chain-map Hamiltonian obtained from the spectral density J defined over support.

Arguments

J::Function: non-negative function defined (at least) on support.

support: support of J; it can be a pair of real numbers, denoting an interval, but can be more generally a list of increasing real numbers [a_1, a_2, a_3, …, a_N], that will be interpreted as $(a_1,a_2) ∪ (a_2,a_3) ∪ ⋯ ∪ (a_{N-1},a_N)$. The subdivision is ignored for the calculation of the chain coefficients, but it can become useful to compute the integral of J over support, since the numerical integration algorithm may need to exclude some intermediate points from the domain to work (e.g. if they are singularities).

L::Int: number of oscillators in the chain.

Nquad::Int: a parameter passed to OrthoPoly, which determines the number of nodes used for the quadrature method when numerical integration is performed.

Keyword arguments are passed on to the OrthoPoly constructor of the set of orthogonal polynomials.

It returns the tuple (Ω,κ,η) containing

the single-site energies Ω,

the coupling coefficients κ between oscillators,

the coupling coefficient η between the first oscillator and the system.

The spectral function J is associated to the recursion coefficients $αₙ$ and $βₙ$ which make up the recursion formula $x πₙ(x) = πₙ₊₁(x) + αₙ πₙ(x) + βₙ πₙ₋₁(x)$ for the monic orthogonal polynomials $\{πₙ\}_{n\in ℕ}$ determined by J. In the formula, $π₋₁$ is the null polynomial. The chain coefficients are then given by

$Ωᵢ = αᵢ$, with $i\in\{1,\dotsc,L\}$,

$κᵢ = \sqrt{βᵢ₊₁}$, with $i\in\{1,\dotsc,L-1\}$,

while η is the integral of J over its support.

TEDOPA variants

TEDOPA.chainmapping_tedopa — Function

chainmapping_tedopa(file::IOStream)

Return the frequency and coupling coefficients of the TEDOPA chain obtained by the environment described in the JSON file file.

See chainmapping_tedopa for more information.

chainmapping_tedopa(filename::AbstractString)

Return the frequency and coupling coefficients of the TEDOPA chain obtained by the environment described in the JSON file called filename.

See chainmapping_tedopa for more information.

chainmapping_tedopa(parameters::Dict{<:AbstractString, Any})

Return the frequency and coupling coefficients of the TEDOPA chain obtained by the environment specified by the parameters dictionary.

Example

julia> env = Dict(
    "domain" => [0, 2],
    "spectral_density_parameters" => [1, 0.5],
    "spectral_density_function" => "sqrt((2a[2]-a[1]+x)*(2a[2]+a[1]-x))",
);
julia> p = Dict(
    "environment" => env,
    "chain_length" => 200,
    "PolyChaos_nquad" => 5000,
);
julia> chainmapping_tedopa(p)
(
    frequencies=[1.0000000000000004, …, 1.000000000000001],
    couplings=[0.22360679783266632, …, 0.49999999955894253],
)

TEDOPA.chainmapping_ttedopa — Function

chainmapping_ttedopa(file::IOStream)

Return the frequency and coupling coefficients of the TEDOPA chain obtained by the environment described in the JSON file file, after a thermalization procedure.

See chainmapping_tedopa for more information.

chainmapping_ttedopa(filename::AbstractString)

Return the frequency and coupling coefficients of the TEDOPA chain obtained by the environment described in the JSON file called filename, after a thermalization procedure.

See chainmapping_tedopa for more information.

chainmapping_ttedopa(parameters::Dict{<:AbstractString, Any})

Return the frequency and coupling coefficients of the TEDOPA chain obtained by the environment specified by the parameters dictionary, after a thermalization procedure.

See chainmapping_tedopa for more information.

TEDOPA.chainmapping_tftedopa — Function

chainmapping_tftedopa(file::IOStream)

Return the frequency and coupling coefficients of the TEDOPA chain obtained by the environment described in the JSON file file, after transforming it into a T=0 and μ=0 environment through the TF-TEDOPA algorithm.

See chainmapping_tedopa for more information.

chainmapping_tftedopa(filename::AbstractString)

Return the frequency and coupling coefficients of the TEDOPA chain obtained by the environment described in the JSON file called filename, after transforming it into a T=0 and μ=0 environment through the TF-TEDOPA algorithm.

See chainmapping_tedopa for more information.

chainmapping_tftedopa(parameters::Dict{<:AbstractString, Any})

Return the frequency and coupling coefficients of the TEDOPA chain obtained by the environment specified by the parameters dictionary, after transforming it into a T=0 and μ=0 environment through the TF-TEDOPA algorithm.

See chainmapping_tedopa for more information.

TEDOPA.chainmapping_thermofield — Function

chainmapping_thermofield(file::IOStream)

Return the frequency and coupling coefficients of the TEDOPA chain obtained by the environment described in the JSON file file, after transforming it into a T=0 and μ=0 environment through the thermofield procedure.

See chainmapping_tedopa for more information.

chainmapping_thermofield(filename::AbstractString)

Return the frequency and coupling coefficients of the TEDOPA chain obtained by the environment described in the JSON file called filename, after transforming it into a T=0 and μ=0 environment through the thermofield procedure.

See chainmapping_tedopa for more information.

chainmapping_thermofield(parameters::Dict{<:AbstractString, Any})

Return the frequency and coupling coefficients of the TEDOPA chain obtained by the environment specified by the parameters dictionary, after transforming it into a T=0 and μ=0 environment through the thermofield procedure.

See chainmapping_tedopa for more information.