Reference

How to describe environments

Each variant of TEDOPA starts from the description of an environment: in this package, information about an environment must be encoded in a dictionary with the following structure: here is a breakdown of all the required keys.

The "environment" key collects the information about the environment, and contains the following sub-keys.

"spectral_density_function" is a string containing a valid Julia function of the variable x, and possibly a set of additional parameters that must be named a[1], a[2], and so on.

"spectral_density_parameters" is a list of numbers that will replace a[1], a[2] etc., in the given order, in the spectral density function. If no additional parameters are needed in the spectral density function, assign the empty list [] to this key.

"domain" is the support of the spectral density function. For the purposes of this package, this is always an interval of the real line, than can be given as a list of two real numbers, representing the endpoints of the interval. If the spectral density function contains singular points, it is strongly advised to add those points to the domain (following the proper numerical order). These intermediate points will be taken into account by the numerical integration routines, leading to more precise results. For example, for a spectral density function defined on $[0,1]$ with singular points in $1/4$ and $1/2$ we should set the domain key to [0,0.25,0.5,1].

"temperature" is the temperature of the environment.

"chemical_potential" is the chemical potential of the environment, which is a number that must lie within the domain of the spectral density function.

The last two items can be omitted if the algorithm doesn't need them.

There are then two additional parameters, which are not related to the original environments.

"chain_length" specifies the lengths of the resulting chains, i.e. how many sites they will have.

"PolyChaos_nquad", an integer, is a parameter that will be passed to the numerical integration routines used to calculate the coefficients (usually a higher number results in a higher precision).

Examples

The following dictionary describes a (simplified) super-Ohmic spectral density on $[0,2]$ with a hard cutoff: $J(x) = x^2$. This can be encoded in a dictionary in the aforementioned format as follows:

Dict(                                 
    "environment" => Dict(
        "spectral_density_parameters" => [],
        "spectral_density_function" => "x^2",
        "domain" => [0, 2],    
    ),
    "chain_length" => 100,
    "PolyChaos_nquad" => 200,
)

This other dictionary describes an environment with a semi-elliptical spectral density

\[J(x) = \frac{1}{2\pi} \sqrt{x(2-x)}\]

on the domain $[0,2]$, and given temperature and chemical potential:

Dict(
    "environment" => {
        "spectral_density_parameters" => [1, 0.5],
        "spectral_density_function" => "1/2pi * sqrt((2a[2]-a[1]+x)*(2a[2]+a[1]-x))",
        "domain" => [0, 2],
        "temperature" => 0.4,
        "chemical_potential" => 0.5
    },
    "chain_length" => 200,
    "PolyChaos_nquad" => 5000,
)

Note that the use of spectral function parameters is not mandatory, and is purely for convenience.

Chain-mapping functions

This package offers several ways of computing the chain mapping, all derived from the original TEDOPA algorithm.

chainmapping_tedopa: standard chain mapping of a single environment;

chainmapping_ttedopa: single-chain thermalised chain mapping for bosonic environments;

chainmapping_tftedopa: single-chain thermalised chain mapping for fermionic environments;

chainmapping_thermofield: a thermofield transformation and then a chain mapping of the resulting (fermionic) environment; it can also merge multiple multiple environments together (in this case, the "environment" subdictionary is replaced by a list of dictionaries, as illustrated in the example below).

The dictionary can be directly supplied as an argument, directly as a Julia Dict object (more precisely of the Dict{<:AbstractString,Any} type) to these functions. Alternatively, the dictionary can be stored in a file, in JSON format; then you can supply to these functions either a string containing the file name, or the file stream object directly.

Example

Given a Julia dictionary dict such as the two ones in the How to describe environments, we can call one of the four methods above with dict as an argument, for example chainmapping_ttedopa(dict). Say we want, instead, to store the information in a JSON file called ohmic.json, with the following contents:

{
    "environment": {
        "spectral_density_parameters": [1, 0.5],
        "spectral_density_function": "1/2pi * sqrt((2a[2]-a[1]+x)*(2a[2]+a[1]-x))",
        "domain": [0, 2],
        "temperature": 0.4,
        "chemical_potential": 0.5
    },
    "chain_length": 200,
    "PolyChaos_nquad": 5000
}

We can then directly call chainmapping_ttedopa("ohmic.json"), or first open the file in Julia and then pass the stream object to the chain-mapping function:

open("ohmic.json", "r") do f
    chainmapping_ttedopa(f)
end

Core algorithm

This is the chain-mapping algorithm, detailed in [2], that lies at the core of the package.

Let $f \colon X \to [0, +\infty)$ be a measurable function, and $\{p_n\}_{n=0}^{+\infty}$ be the set of monic orthogonal polynomials with respect to the measure $\dd f(x) = f(x)\,\dd x$. They satisfy

\[x p_n(x) = p_{n+1}(x) + \alpha_n p_n(x) + \beta_n p_{n-1}(x)\]

for all $n \in \N$, with $p_{-1} \equiv 0$.

The (internal) chainmapping method, given the function $f$, returns the sequences of the recurrence coefficients $\alpha_n$ and $\beta_n$. Remember that, for the purposes of this package, we always assume that the support of $f$ is an interval on the real line.

Standard TEDOPA chain mapping

The chain-mapping algorithm is used by TEDOPA in order to transform the original continuous environment into a discrete chain of modes. Consider an environment described by the free Hamiltonian

\[H\env = \int_X x \adj{f_x} f_x \,\dd x\]

where $f_x$ and $\adj f_x$ are the annihilation and creation operators, respectively, of the environment mode at frequency/energy $x$, and, for example, the interaction Hamiltonian

\[H\inter = \int_X (\adj{A\sys} f_x + \adj{f_x} A\sys) \sqrt{J(x)} \,\dd x\]

although other interaction forms can be used, as long as the operator is linear in $f_x$ and $\adj f_x$. Then, the unitary transformation

\[b_n = \int_X \conj{u_n(x)} f_x \,\dd x, \qquad u_n(x) = \sqrt{J(x)} \frac{p_n(x)}{\norm{p_n}}\]

generates a new set of canonical operators $b_n$ that satisfy the same statistics (bosonic or fermionic) as the original $f_x$ operators. In terms of the new mode operators,

\[H\env = \sum_{n=0}^{+\infty} \bigl[\omega_n \adj{b_n} b_n + \kappa_n (\adj{b_n} b_{n+1} + \adj{b_{n+1}} b_n) \bigr],\]

and

\[H\inter = \eta (\adj{A\sys} b_0 + \adj{b_0} A\sys)\]

where $\omega_n = \alpha_n$, $\kappa_n = \sqrt{\beta_{n+1}}$ for $n \in \N$ and $\eta = \norm{p_0}$, where the monic orthogonal polynomials and the recurrence coefficients are taken with respect to the measure induced by $J$. The chainmapping_tedopa method takes the spectral density function $J$ and computes the frequencies $\omega_n$, the couplings $\kappa_n$ and the system-environment coupling $\eta$.

TEDOPA.chainmapping_tedopa — Function

chainmapping_tedopa(parameters::Dict{<:AbstractString,Any})
chainmapping_tedopa(file::IOStream)
chainmapping_tedopa(filename::AbstractString)

Compute the frequency and coupling coefficients of the TEDOPA chain obtained by the environment specified by either the parameters dictionary or by an external JSON file, passed as a stream object file or by its name filename.

Return a struct whose data can be retrieved via the following methods:

frequencies for the single-site frquencies/energies;

couplings for the coupling coefficients (the first element of the list is the coupling coefficient between the system and the first site of the chain);

domain for the domain of the spectral density function;

the spectral density function can be accessed by calling the struct directly, as a function.

Example

julia> dict = Dict(
    "environment" => Dict(
        "spectral_density_parameters" => [],
        "spectral_density_function" => "x^2",
        "domain" => [0, 2],
    ),
    "chain_length" => 100,
    "PolyChaos_nquad" => 200,
);

julia> env = chainmapping_tedopa(dict);

julia> frequencies(env)
100-element Vector{Float64}:
 1.4999999999999993
 1.1666666666666665
 1.083333333333333
 1.0499999999999998
 1.0333333333333332
 ⋮
 1.0001073883161524
 1.0001051967178594
 1.0001030715316432
 1.0001010101010104
 1.0000990099009903

julia> couplings(env)
100-element Vector{Float64}:
 1.632993161855452
 0.3872983346207419
 0.45074893585520914
 0.47245559126153397
 0.48241815132442156
 ⋮
 0.49995252761390824
 0.4999535013925803
 0.4999544455135774
 0.49995536116913225
 0.49995624949217776

julia> domain(env)
2-element Vector{Float64}:
 0.0
 2.0

julia> env(2)
4

The function returns a struct env that contains the following data.

A list of the chain mode frequencies $\omega_n$, that can be accessed with frequencies(env).

A list of the chain mode couplings, whose first element is $\eta$ followed by the $\kappa_n$ coupling coefficients, and that can be retrieved with couplings(env).

The domain of the spectral density function, domain(env).

The spectral density function, that can be accessed simply by calling env itself with a real number as an argument.

The last two features are clearly useless in this case, as we already know the spectral density function from the beginning, but they can be useful with the other TEDOPA methods that transform the function.

Thermalised TEDOPA

When the environment starts from a thermal state at inverse temperature $\beta$ and the interaction is of the form

\[H\inter = \int_X A\sys (f_x + \adj{f_x}) \sqrt{J(x)} \,\dd x\]

then the thermalised TEDOPA approach may be used, in which we replace the original environment with an equivalent one with the following features.

The new environment starts from the vacuum state.

Assuming the original domain is $[0,x\Max]$, the new environment is described by a continuum of modes $g_x$ (following the same statistics as the $f_x$ modes) with frequency/energy $x \in [-x\Max, x\Max]$ and Hamiltonian operators

\[H\env' = \int_{-x\Max}^{x\Max} x \adj{g_x} g_x \,\dd x\]

and

\[H\inter' = \int_{-x\Max}^{x\Max} A\sys (g_x + \adj{g_x}) \sqrt{J'(x)} \,\dd x,\]

where $J'$ is the thermalised spectral density function

\[J'(x) = \frac12 \sgn x\, J(\abs{x}) \Bigl( 1 + \coth\frac{\beta x}{2} \Bigr)\]

for a bosonic environment, or

\[J'(x) = \frac12 J(\abs{x}) \Bigl( 1 + \tanh\frac{\beta x}{2} \Bigr)\]

for a fermionic one.

The chainmapping_ttedopa and chainmapping_tftedopa methods, for bosonic and fermionic environments respectively, return the chain coefficients and the modified spectral density function of the thermalised environment.

TEDOPA.chainmapping_ttedopa — Function

chainmapping_ttedopa(parameters::Dict{<:AbstractString, Any})
chainmapping_ttedopa(file::IOStream)
chainmapping_ttedopa(filename::AbstractString)

Return the thermalised spectral density and the associated frequency and coupling coefficients obtained by the T-TEDOPA transformation of the environment specified by either the parameters dictionary or by an external JSON file, passed as a stream object file or by its name filename.

See chainmapping_tedopa for more information.

Example

julia> dict = Dict(
             "environment" => Dict(
                 "spectral_density_parameters" => [],
                 "spectral_density_function" => "sqrt(x) * exp(-x/0.5)",
                 "domain" => [0, 2],
                 "temperature" => 2.5,
             ),
             "chain_length" => 100,
             "PolyChaos_nquad" => 200,
         );

julia> env = chainmapping_ttedopa(dict);

julia> frequencies(env)
100-element Vector{Float64}:
  2.7937503875293512e-8
  0.2377475766338636
 -0.06307177633684016
  0.07702453114697957
 -0.07951306693183548
  ⋮
  0.006497663318373198
 -0.00645973925374582
  0.006422260704823025
 -0.006387253852889501
  0.006352720782309697

julia> couplings(env)
100-element Vector{Float64}:
 2.500086760123663
 0.000376728898182609
 1.07931125845791
 0.9389853648435562
 1.1159207167585214
 ⋮
 0.9923014215388832
 1.0076699809013268
 0.9923919032035641
 1.0075820962595274
 0.9924754152732462

julia> domain(env)
3-element Vector{Float64}:
 -2.0
 -0.0
  2.0

julia> env(1.0)
0.410505041660018

TEDOPA.chainmapping_tftedopa — Function

chainmapping_tftedopa(parameters::Dict{<:AbstractString, Any})
chainmapping_tftedopa(file::IOStream)
chainmapping_tftedopa(filename::AbstractString)

Return the thermalised spectral density and the associated frequency and coupling coefficients obtained by the TF-TEDOPA transformation of the environment specified by either the parameters dictionary or by an external JSON file, passed as a stream object file or by its name filename.

See chainmapping_tedopa for more information.

Example

julia> dict = Dict(
             "environment" => Dict(
                 "spectral_density_parameters" => [],
                 "spectral_density_function" => "1/pi * sqrt(x*(2-x))",
                 "domain" => [0, 2],
                 "temperature" => 10.0,
                 "chemical_potential" => 1,
             ),
             "chain_length" => 100,
             "PolyChaos_nquad" => 200,
         );

julia> env = chainmapping_tftedopa(dict);

julia> frequencies(env)
100-element Vector{Float64}:
  0.012494792328180852
  5.2026993957315084e-6
 -1.077894557975692e-9
 -1.7277946434351792e-9
 -1.7269887048876722e-9
  ⋮
  7.564160055262281e-9
  8.971713063438136e-9
  1.0968847909825774e-8
  1.4329892390751997e-8
  2.447297564427131e-8

julia> couplings(env)
100-element Vector{Float64}:
 0.707106781448028
 0.499843803889265
 0.4999998810351929
 0.49999987899169407
 0.49999984443772305
 ⋮
 0.4999988694272515
 0.4999990207121836
 0.49999920014918586
 0.4999994195321602
 0.4999997061585101

julia> domain(env)
2-element Vector{Float64}:
 -1.0
  1.0

julia> env(-0.2)
0.308760037243951

Thermofield+TEDOPA (for fermionic environments)

In the more common case of fermionic environment with an exchange-type interaction, the above thermalisation procedure is not applicable. There is another approach, based on the thermofield transformation: we introduce a second, ancillary environment, then we shift the dependence of the initial state of the environment on the temperature and the chemical potential to the coupling between the system and the two environments. Namely, we transform

\[H\env = \int_X (x-\mu) \adj{f_x} f_x \,\dd x,\\ H\inter = \int_X (\adj{A\sys} f_x + \adj{f_x} A\sys) \sqrt{J(x)} \,\dd x\]

into

\[H\env' = \int_X x (\adj{f_{0,x}} f_{0,x} + \adj{f_{1,x}} f_{1,x}) \,\dd x,\\ H\inter' = \int_X (\adj{A\sys} f_{0,x} + \adj{f_{0,x}} A\sys) \sqrt{J_0(x)} \,\dd x + \int_X (\adj{A\sys} f_{1,x} + \adj{f_{1,x}} A\sys) \sqrt{J_1(x)} \,\dd x\]

with the two new spectral density functions

\[J_0(x) = \biggl( 1 + \frac{1}{e^{\beta(x-\mu)} + 1} \biggr) J(x),\\ J_1(x) = \frac{1}{e^{\beta(x-\mu)} + 1} J(x).\]

The $f_{0,x}$ modes will start from the empty state, while the $f_{1,x}$ ones will start from the filled state. The chainmapping_thermofield method returns a named tuple with two fields, empty and filled, each one referring to the chain coefficients and the modified spectral density function of the thermalised environment starting from the empty or filled state, respectively.

Multiple initial environments

The thermofield+TEDOPA method can be also be applied when the system interacts with multiple environments at once. In order to use the algorithm in this situation, simply replace the "environment" dictionary with a list of dictionaries, one for each individual environment, as in the following example.

{
    "environment": [
        {
            "spectral_density_parameters": [],
            "spectral_density_function": "1/(pi*200)",
            "domain": [-100, 100],
            "temperature": 0,
            "chemical_potential": 0.01
        },
        {
            "spectral_density_parameters": [],
            "spectral_density_function": "1/(pi*200)",
            "domain": [-100, 100],
            "temperature": 0,
            "chemical_potential": -0.01
        }
    ],
    "chain_length": 200,
    "PolyChaos_nquad": 5000
}

TEDOPA.chainmapping_thermofield — Function

chainmapping_thermofield(parameters::Dict{<:AbstractString, Any})
chainmapping_thermofield(file::IOStream)
chainmapping_thermofield(filename::AbstractString)

Return the thermalised spectral densities and the associated frequency and coupling coefficients obtained by the thermofield+TEDOPA transformation of the environment specified by either the parameters dictionary or by an external JSON file, passed as a stream object file or by its name filename. The returned object is a NamedTuple with two fields, empty and filled: the former contains the information relative to the final environment that starts from the vacuum state, the latter containes information about the completely filled state.

See chainmapping_tedopa for more information.

Example

julia> dict = Dict(
             "environment" => Dict(
                 "spectral_density_parameters" => [],
                 "spectral_density_function" => "1/2pi * sqrt(x*(2-x))",
                 "domain" => [0, 2],
                 "temperature" => 10.0,
                 "chemical_potential" => 0.5,
             ),
             "chain_length" => 50,
             "PolyChaos_nquad" => 200,
         );

julia> env = chainmapping_thermofield(dict);

julia> frequencies(env.empty)
50-element Vector{Float64}:
 1.0121826824869755
 1.0000050727401346
 0.999999998949
 0.9999999983153369
 0.9999999983161233
 ⋮
 0.9999999985991959
 0.9999999986176314
 0.9999999986371221
 0.9999999986577404
 0.9999999986795602

julia> couplings(env.filled)
50-element Vector{Float64}:
 0.34910972608058455
 0.49983990373980103
 0.4999998802225264
 0.4999998789916133
 0.49999984443770995
 ⋮
 0.4999984995521281
 0.4999984708186977
 0.49999844246334957
 0.49999841450784854
 0.4999983869753035

julia> domain(env.filled)
2-element Vector{Float64}:
 0.0
 2.0

julia> env.empty(0.2)
0.04703033939361449

julia> env.filled(0.2)
0.04846262646152273